consume air quality API and manage cache
Retrofit was utilized to obtain API responses from GIOS public air
quality API. Repositories were implemented to cache data and correctly
choose when to reach out to the server for an update and when to use
cached copies of older data.

MainActivity, at least temporarily can now access the air quality data
and display it, as well as request an update. The way the data is
displayed it kept very basic for now.