fix crash on internet connection unavailable
Surrounded retrofit calls execution with try-catch blocks. Also adjusted
update policy by adding a new minimum interval value after a failure to
connect and retrieve a list of air quality measurement stations when
there are no stations cached - in this case the interval is set to 30
minutes, and matches the regulations of the API access.

Addidionally changed the minium interval between air quality readings
download, to match the same regulations.