limit accepted distance from station to 10 km
Stations farther than 10 km away are no longer accepted as air quality
data sources. If no station is within the 10 km radius, data is not
downloaded at all (other than periodic stations cache updates) and an
AirQualityLog is saved with a new error code indicating this scenario
being the reason for the lack of data.