code cleanup
Renamed property indexRaw in AirQualityLog class to AirQualityIndex, and
changed the database column name for it to air_quality_index. This is to
make the property more obvious as to what it is.

Removed unused value in MainActivityViewModel.

Renamed various properties related to GIOS API responses as they were
foreign to the apps naming policy, like "gegrLon" instead of "longitude"
or "stIndexLevel" instead of "indexLevel".