introduce a cleaner-to-use parameter 'sensitivity'
Previously when setting the desired minimumWarningIndexLevel,
MainActivity had to calculate the desired airQualityIndex itself,
because it's dependent on the UI used (how are different levels of
the index expressed). Now a 'sensitivity' parameter has replaced the
minimumWarningIndexLevel and the calculation between scales is done once
in a single place (in the AirCheckParams class).

Sensitivity setting describes how sensitive the app is to smog in terms
of how bad the air pollution has to be in order to trigger a warning.