improve user experience with settings
Each level of sensitivity setting, regarding how bad the smog has to be
in order to trigger a warning from the app, is now explained to the user
right above the Seeker used to change the level.

Additionally increased contrast, introducing an amoled-friendly black
background, in line with the app's battery-friendly philosophy.