fix a rare issue preventing AQ updates
While unlikely it is still possible that a user would keep the app open
for hours straight while continously changing the sensitivity setting.
Since the change reschedules the AQ update job, such user behavior would
effectively stop the updates altogether. Thus if a user is waiting for
the AQ change while staying in the app and plays with the setting, they
would never see any new data.

This is now fixed by attempting an AQ update every time the setting is
changed, therefore the AQ is checked both immediately upon the setting
change, and the next update is rescheduled for a sensible interval
ahead.

This is efficient, because AirQualityLogsRepository will use
cached value unless it makes sense to download a new one. And in case it
does make sense to fetch new data - it's welcome. File IO is unlikely
too as SQLite is used for storing the cached api responses as well as
they are likely to be fetched before and kept in memory already.
Otherwise they will be fetched from storage just once, the first time
the setting is changed.

The update is ran on a background thread.