implement notification choice logic
Depending on the two latest AirQualityLogs a AQLogsComparer is now
deciding whether or not to display a notification about air quality
change. It also decides when to show an error notification or a lack of
data notification.

NotificationHelper was modified, new notification channels were added,
so the users of Android 8+ can choose which types of notifications when
want to receive.

Tests were added to ensure intended behaviour on the part of the
AQLogsComparer.