fix potential issue with wakeLock
When JobService's onStartJob returns false, the system keeps a wakeLock
for it till jobFinished() is called. In AirQualityCheckJobService
AsyncTask is executed in onStartJob then false is returned. However the
task is referenced with a nullable variable, so task?.execute() notation
was used. This will not start the task if the variable is null, but
later false is returned anyway. Thus if the task is null somehow, the
wakeLock will be kept for a very long time unnecessarily, as the taks is
calling the jobFinished().

This was fixed by adding an elvis operator and returning true if the
variable was null.