add tests for StationsRepository
The repository is crucial for the app's proper functioning, as without
the stations list, no data can be retrieved about air quality readings.

The repo must correctly choose when to update cached stations list and
when to reach out to the API for a fresh list of stations.
It must detect changes between cache and API's response and then update
cache as needed.