fix showing notification for already seen data
When user opens the app, it might fetch fresh data from API and show it
to the user. In such a case it's undesirable to show the user a
notification about the data they've just seen a couple minutes later.

This is now fixed by both requiring that the data is fresh before pushing
notifications, and rescheduling the next update after
MainActivityViewmodel gets fresh data.