harden the app against spotty and failing stations
Some of the air pollution measurement stations are of lower quality than
others. This may manifest in their failing to provide AQ index at all,
or by either temporarily or consistently outputing deficient data.

In order to harden the app against those stations, it's been allowed to
try a second station in case the nearest one responds with low quality
data or lack of AQ index entirely. If the second nearest station turns
out to have better quality data and it's AQ index is worse than the one
returned by the nearest station, the further station's data is used.

AirQualityLog now includes data about sensors used in computing the AQ
index value it holds. This is kept in an efficient FLAG-based manner.