allow more frequent update attempts
When an update attempt fails due to an error like no internet access or
no location access, there is no need to restrict further attempts in
near future. The restriction was instilled in order to enforce the
regulations of API access, but when no actual requests were sent to the
server, then there is no point in refusing to make another attempt when
the user opens the app again, for example, after fixing the issue.

AirQualityLogsRepository now allows more frequent updates unless it had
connected the API during the previous update.