efficiently encode air quality details for storage
PollutionDetails class was added, which stores sub-indexes for all
pollutants measures. It encodes the data for storage purposes, where it
takes a single integer to hold all 7 currently existing types of
pollution. In case sensor range increases in the future, there is a lot
of headroom, particularily by moving over to long type. This is however
very unlikely to be needed and some more senors can be taken care of
without the move too.

A separate class for storing covered sensors was added, providing
efficient, flag-based way of checking which sensors have been used.
PollutionDetails return their data about covered sensors in this form.

Tests were added to cover the encoding/decoding of subIndex data, as
this is crucial for the proper functioning of the app.

PollutionDetails is now part of the AirQualityLog, as details, and is
saved/retrieved by Room via a custom TypeConverter.

Older tests were updated to keep up with the changes.