let StationsRepository fetch and cache sensors
Available sensors per station are not cheap to obtain beforehand, as
this would require more than 100 API calls in quick succession. However,
for most users only sensors of a couple of stations are of any
importance, which allows for a much more efficient solution. The
solution is to only ask for the sensors of stations that were actually
encountered, that is a user was in their vicinity physically.

Also updated StationsRepositoryTest, with new signature and a new mock
class.