change cache metadata storage mechanism
Periodically data about air quality measurement stations is updated in a
local cache. Data about the last update time and status used to be
stored with a custom, SharedfPreferences-based solution. Since this
setup required keeping the cusrom code in sync with Room database,
problems could arise when database is updated and the custom metadata is
not properly modified during the migration.

Therefore the old solution was removed and a new table was added to Room
database, holding logs of cache updates. StationsRepository now uses
this table's dao instead of the custom solution.