handle partial data availability
When deciding whether or not to notify the user about either data
shortage, bad or good air quality, several factors need to be
considered. With the addition of partial data handling in the app, the
decision making had to be upgraded to deal with these cases well.

To ensure compliance with specification, an exhaustive unit test was
added traversing all possible combinations of previous and current
AirQualityLogs.