delete stations hesitantly, after long absence
Since API response to get all stations request is sometimes missing
stations, like when they experience a major downtime probably due to
maintenance, this change was added. Stations are no longer removed
immediately upon not appearing in the API response. Instead an absence
count is increased, until it reaches threshold value and a given station
is deleted from cache.

An added benefit is that when a sensor type is removed from the area,
the app will show "unknown" index for some time, until it agrees to
delete the station.