use lastLocation unless older than 20 minutes
When deciding whether or not to fetch location actively, LocationHelper
is now checking if the lastLocation is older than 10 minutes, this isn't
perfect, but with some adjustments such an approach might prove
beneficial.

Also adding a flag to AirQualityLog's metadata, indicating that active
location method was utilised whenever it is the case.