remove unnecessary location availability check
In MainActivityViewModel there was a check which used to make sense back
in times when AirQualityLogsRepository did not discriminate between AQ
requests that ended up using AQ API and those which didn't reach API (this
has to do with regulations of API use, max 2 uses per hour) and
prohibited further attempts after failure due to unavailable location
data. Since currently the repository allows consecutive attempts as long
as they're not reaching air quality API, it makes no sense to make this
check in the viewModel.

The check and a dependency it required were removed from the viewModel.
An added benefit is that correct error message will be shown without
doubling the logic of choosing the right error message for the case.