remove busy waiting in active location request
Busy waiting has been replaced with concurrency primitives. Since use of the primitives is generally discouraged, more abstract solutions were examined and experimentally implemented, like ReentrantLock, but ultimately - as none of their extended capabilities were found useful in this simple case - the synchronized blocks were chosen as more elegant and readable than the try-finally blocks used with the ReentrantLock, along with separately declared Condition objects used with it.

CountdownLatch-based solution was rejected due to effectively locking on a mutable entity unless the entire implementation of the working code was changed around the latch.

The chosen implementation explicitly handles spurious wakeups, but it is a minor issue given how widely known the problem is, this would also be the case with ReentrantLock too, as it's not much more abstract. Other than that, the use case seems basic enough to justify use of concurrency primitives, as an exception from the general rule.