add triggers on database open, not create
Room's onCreate callback is only called once per database lifetime, and
is not called during migration, even a destructive migration.
Therefore during a destructive migration, when all triggers are wiped,
Room does not recreate them. This imposes adding a migration where the
triggers are recreated, but since it is a design choice for this app to
not use migrations and instead wipe the database if schema is changed,
this solution was rejected.

Instead onOpen callback is used and triggers are now created with "IF
NOT EXISTS" condition. This solution is acceptable because Room will
only call the callback when db is accessed for the first time after the
app is opened, and the access is made on the same thread as the usage of
the database, which means the SQL is not executed on the UI thread
(because the app is configured so it does not allow that anyway).

This is therefore considered a next-to-nothing performance cost, but at
the same time it simplifies the maintenance and hardens the app against
human error in the future.