code cleanup, minor bugfix and library update
Renamed method in AirQualityLogsRepository to better fit it's return
type.

Translated a comment mistakenly written in Polish.

Added a new build flavour for field testing builds.

Updated location library to version 16.0.0.

Made constants in ActiveLocationRequestHelper private as they don't need
to be accessed from the outside.

Also changed default value of visibility for progressBar in
MainActivity, as it was mistakenly attributes with "tools:" prefix
instead of "android:" one, so the value was unused in actual builds.
This change makes no difference to the user experience as the value was
overwritten immediately on startup anyway, but it still secures the app
against a potential bug in the future versions.