fix incorrect sensors coverage comparison
SensorsPresence class had a bug in it's hasSensors implementation, which
only worked reliably for comparisons with a single flag, and not with a
bunch of flags at once, all kept on a single Int.

This was corrected and tests were added, where the SensorsPresence is
tested against a naive, alternative implementation. Tests are
exhaustive, meaning all possible combinations of flags are covered.