synchronize getLatestLogData() function
The function is supposed to provide the latest data available, given the
restraints of API usage regulations applicable. It should not be called
by two threads at the same time. Therefore the second thread will have
to wait for the first one to finish, so only one set of network requests
will be executed.