test AirQualityLogsRepository and fix bugs found
Along added Mockito library and used it to mock the repo's dependencies.

Mockito was set to allow mocking final classes by the addition of
the resources/mockito-extensions/org.mockito.plugins.MockMaker file.

Immediately it's become obvious that there is a serious bug in the
fusion mechanism in the repo. In cases of data shortage the mechanism
wasn't trying to find missing data on the subsequent stations if they've
provided no other sensors than the ones expected so far. Thus the fact
the data is missing was ignored in some cases.
The bug was hard to notice in real life, also hard to tell apart from a
simple data shortage, which isn't a bug in the app but a problem on the
API side. This is now fixed by using gained coverage instead of the
expected coverage as a comparison point. Essentially this means that if
the considered station has sensors that aren't yet covered, it will be
asked for the values. This behaviour was intended all along.