remove error message logic from viewModel
MainActivityViewModel used to have a direct access to UI via a Toast
with error message. The error covered by it is expected to never occur
anyway, due to it's character (occurs when jobScheduler is getting an
invalid job to schedule or when the app doesn't provide it a correct
service).

ViewModel should not decide how UI behaves and how data is shown to the
user. Therefore the error message was removed and not replaced with any
ther mechanism, as it's value to the user is limited and it's occurrence
not plausible. Additionally failure of the job scheduling process is
visible to the user anyway, as the setting does not change, which is
much more clear than a cryptic error message.