test StationDataConverter and fix potential crash
Added unitTest for the aforementioned class and noticed a possible crash
when API responds with malformed stations list. In such a case the app
should not crash, so try-catch block was used. There are potential
issues with reuse of this code - for example if many stations' data is
malformed in API's response - when iterating through the objects
returned by the API and throwing exceptions for each of them, some
performance price would have to be payed for it. Still, it's better not
to crash than it is to avoid a performance price. Both mistakes seem
similarily likely, so the least grave was chosen.