test SensorsDataConverter and fix issues found
Tests were added for the class and two minor issues were found:

I. The return type of the toSensorFlags() was Int? but the function has
no capacity to return null anyway, so this was leading to mistaken
assumptions in usage.

II. Such a mistaken assumption was fixed in StationsRepository, as it
assumed that null might be returned and was ready to act upon that. It
seemed like perhaps it's an answer to the case of no flags returned,
which made great deal of sense so it was reimplemented to actually work
like that. This means that when no flags are returned by the function,
StationsRepository now correctly refuses to update sensors value in the
database (because it has to be zero anyway in that case, so it wouldn't
change) instead it returns null indicating correctly that it has found
nothing. AirQualityLogsRepository uses this null response to prevent
downloading data from the API in case no sensors are found, so the bug -
although rare and unexpected as API should not be expected to return 0
sensors for a station(why would a station with no sensors even exist?)
still, anything can happen on the server-side and there is no control
over it.