fix wakelock not being engaged correctly
The Boolean returned in onStartJob indicates whether or not a wakelock
should be held for the service, which is the case when work is still
being done on a separate thread with AsyncTask.

The return value was set incorrectly, that is when the task ran it was
set to false and when the task turned up to be null, and thus could not
run, it was true (indicating the wakelock should be held). This was
reversed in order to fix the issue.