code cleanup
Separated tests for AirQualityLogsComparer into two classes based on the
types of the tests.

Improved the readability of the test output as well as that of the code
itself.

Cleaned up code in AirQualityCheckJobService to be easier to read,
removed an unnecessary null check and added a comment explaining the
purpose and rationale behind the code which ends the job when the latest
log is coming from cache.