introduce automatic retry on error
Instead of immediately showing an error, the app is going to schedule a
second one time job to re-check if the error occurs again. This is
useful because most of theese errors are random and insignificant, like
a temporary internet connection issues or location inavailability.

Should the error persist after the retry delay, which is less than the
minimum delay till the next periodic update, the error will be shown to
the user via a notification.