fix stations missing right after cache update
StationsRepository was returning the list from API when finished with
updating it's cache. This lead to missing the stations that were
absent temporarily. The issue only happened right after cache update.

This was fixed by returning the stations from dao instead, to make sure
all updates are correctly applied.

Also added a test failing with this bug to prevent regression, and fixed
older tests, which relied on the old implementation.