explain partial data better
When only partial data is available, the app now makes a prediction what
should be expected to contribute to the smog issue the most at the
moment. In order to achieve that current month of year number is
obtained and used against a predefined expectations regarding that time
of year. For example in the Winter, particulate matter is required to be
covered, unless no such sensors are present in the area anyway, in which
case the app will assume that it isn't required.

The procedure is using Locale.US for compatibility reasons and because a
small shift in hours, and thus sometimes month numbers, is acceptable.