address accidental location permission denial
When the app does not have the permission to use location, or location
is off on the device or when GooglePlayServices's version on the device
is too old, MainActivity will repeatedly - with delay between - show
dialogs allowing user to immediately fix any and all of these issues
with a single click each.

This handles the situation where the user accidentally denies at first,
but then realizes the mistake. A new dialog will be shown shortly.