require seasonal key pollutants coverage
AirQualityLogsRepository now expands it's expectations regarding sensors
coverage by seasonal key pollutants. When the minimum requirement is not
met, the log's index is -1 even if all the sensors present on the used
measurement stations were covered.

For example if O3 sub-index is missing in the middle of Summer, but the
only nearby station is not meant to provide the sub-index, then from now
on the repo will assing -1 overall index to the log.