add test for seasonal key pollutants handling
AirQualityLogsRepository is expected to include seasonal key pollutants
in the expected sensor coverage of an AirQualityLog. Therefore it checks
if the sensors are covered after all the data has been downloaded from
API, and if an essential pollutant was not expected based on sensor
ranges on the nearby stations, it is added anyway, so the AirQualityLog
contains information about the fact that not all of the necessary
pollutants are covered.
This is now addressed by a simple test to guard against regressions
here.