notify about improvement during partial shortage
Before, when data shortage occured, it used to no longer push
improvement notifications untill the shortage ended entirely. This new
implementation changes how the log is created in the first place: when
seasonal key pollutants are present, the log is having an index even if
some data is missing. The log is later processed, for the most part, as
if it had all the expected data.
Such a partial log is now presented with a comment about it's partial
nature, but the index level is shown normally. Also on the notifications
it is mentioned that the improvement is "likely" not sure. This very
correctly explains the nature of the log.

Alert notifications are unaffected, as they always show when something
is past the threshold.

"Data shortage ended" notifications are shown after everything is
alright and all the expected data is available.