fix shortage notification being shown needlessly
When air quality is bad and a new log comes with bad air as well, but
only partial data, the log is no longer triggering a notification about
data shortage. This is because such a notification is supposed to let
user know that the air might possibly be good but there is no data. When
it's known that the air is still bad, there is no reason to bother the
user.