fix continous errors on physical station removal
When a station is removed from the system and API no longer returns it's
sensors, the app no longer considers lack of the sensors data an error,
instead it assumes safely that the station must have had all the
sensors, and thus the expected coverage is raised. If other stations can
fill all the data in, then no error is needed and results are presented
normally; else some form of a shortage or error is shown.