introduce timeout mechanism for API calls
While individual calls have their own timeouts, there might be many of
such calls executed in succession, in such a case a global timeout
mechanism is necessary to ensure a reasonable overall timeout.

Such a mechanism was added and global timeout was set to 50 seconds.
Notice that it only is checked between API calls, so overall max time is
longer by a single call's timeout.