fix settings being lost on Android update
Settings saved as PersistableBundle via JobScheduler might be lost due
to significant system-wide events, like operating system update. In
order to restore the settings and reschedule the job, the app now saves
a backup of settings and validates the active settings at multiple
check points. If settings bundled with a scheduled job are missing (for
example the job doesn't exist at all) job is scheduled with correct
settings anew.

In order to achieve that the app now implements a
BootFinishedBroadcastReceiver, which takes care of rescheduling the job
on reboot. The periodic background job is no longer setup as persistable,
because the app is doing this job anyway now, so there is no need for the
system to double that functionality.

Validation occurs on boot and on app startup.

Additionally enabled autoBackup, since the settings are now kept within
it's reach.