manually control wakelocks during update
JobScheduler is keeping a wakelock for a job being done, however it
releases it onStopJob, which makes the app misbehave in rare cases. The
intended behavior is to continue work if it was started and handle
errors consistently, with the same code, as if they occured without
jobScheduler's cancellation of the job.

Therefore the wakelock offered by the jobScheduler is rejected and not kept
for the execution time, instead AsyncTask is keeping it's own wakelock.