fix tests failing with coroutines
Coroutines used to ran after the tests' assertions, thus skewing
results.

In order to straighten it out, the tested method with inner coroutine is
now returning a Job, on which a .join() is called in the tests, so they
wait. In order to call join() on a coroutine Job, the call must be
confined within a coroutine scope, runBlocking is used for that purpose.

Additionally Main dispatcher has to be replaced in order to work in
tests, because it's not available outside of the Android platform, like
in local tests. A piece of code with before and after annotated methods
is used in order to set the replacement up properly.

Also removed a call to the method with coroutines within it from another
method in the tested class - this prevents unwanted calls to the
coroutines which caused exceptions showing up after the tests finished,
albeit without errors. Still, in order to avoid confusion, the logic of
the app was a bit altered: now view calls the method explicitly and the
other method does not contain the call in it.